fix(REM-19201, Core) address code review comment about restoring `_metric`
[see pull request #52](https://gitpub.rakuten-it.com/projects/ECO/repos/alpha-android-perftracking/pull-requests/52/overview)

In case the `Sender` fails somewhere mid-loop of it`s `send(int, int)`
method it may have overwritten the `_metric` field or changed its content.
After failing in `send(int,int)` `SenderThread` will try to send the same
index range of measurement in the buffer again - so we need to cache  the
value of `_metric` at the start or `send(int,int)` and restore it in case
of failure.