fix(REM-21165) work around index arithmetic in corner case of full buffer
### Problem

currently the buffer has a capacity of 512 and the sender will send
every 10 seconds (in the happy case). If the buffer is filled up
within those 10 seconds then the `startIndex` for sending will be equal
to the `endIndex` because the buffer is a ring.
In that case the sender will not send any measurements, because it
"thinks" there are 0 measurements to send (`count == 0`) while actually
there are 512 - and the buffer will not be freed for further measurements.

### Solution

To work around this issue I decided to "fix" the end index to be
`startIndex - 1` - which will allow `Sender` to send & free up to 511
measurements. This also adjusts the `endIndex` in `send(startIndex,endIndex)`
so that the sending loop still works.