refactor(REM-21165): rewrite bugfix based on code review suggestion by petr
see https://gitpub.rakuten-it.com/projects/ECO/repos/alpha-android-perftracking/pull-requests/58/overview

### Problem

currently the buffer has a capacity of 512 and the sender will send
every 10 seconds (in the happy case). If the buffer is filled up
within those 10 seconds then the `startIndex` for sending will be equal
to the `endIndex` because the buffer is a ring.
In that case the sender will not send any measurements, because it
"thinks" there are 0 measurements to send (`count == 0`) while actually
there are 512 - and the buffer will not be freed for further measurements.

### Solution

Refactor `Sender#send(startIndex, endIndex)` to `Sender#send(startIndex, count)`
and pass `MeasurementBuffer.SIZE` as `count` if the buffer is full.